Simplify the deep link workflow, and iOS can share the same code path.
