--- native-apps/unity_ii_deeplink/unity_project/Assets/Android/Plugins/com/icgamekit.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 22bfb1d087e132d49aee8d6b77c16596
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
