--- native-apps/unity_ii_deeplink/unity_project/Assets/Android/Plugins/com/icgamekit/plugin.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3649a729052c9154f92028141feef64c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
