--- native-apps/unity_ii_deeplink/unity_project/Assets/Android/Plugins/com/icgamekit/plugin/ICGameKitPlugin.java.meta
@@ -1,32 +1,0 @@
-fileFormatVersion: 2
-guid: 7e6eb76733b18a441bc73656627d8f52
-PluginImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  iconMap: {}
-  executionOrder: {}
-  defineConstraints: []
-  isPreloaded: 0
-  isOverridable: 0
-  isExplicitlyReferenced: 0
-  validateReferences: 1
-  platformData:
-  - first:
-      Android: Android
-    second:
-      enabled: 1
-      settings: {}
-  - first:
-      Any: 
-    second:
-      enabled: 0
-      settings: {}
-  - first:
-      Editor: Editor
-    second:
-      enabled: 0
-      settings:
-        DefaultValueInitialized: true
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
